requirements.txt
python-socketio
pytest